grpc-mate-python/requirements.txt
-i https://pypi.org/simple
grpcio-tools==1.24.3
grpcio==1.24.3
protobuf==3.10.0
pyyaml==5.1.2
six==1.12.0
sqlalchemy==1.3.10
